Temporarily move mocke to commons-test
